+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Next Version (prediction: 3.0.0)
 
 * **[BC, IMPROVED]** Added team information to player statistics in `BoxScoreResponse`
+* **[NEW]** Added `StatisticsCollection::findGroupByKey()`
 
 ### 2.0.0 (2015-10-15)